Move some timeline related code to the actual timeline class
