+++ src/MaterialEditor.Base/MaterialAPI.cs
@@ -14,8 +14,6 @@ namespace MaterialEditorAPI
         /// Postfix added to the name of a material when copied
         /// </summary>
         public const string MaterialCopyPostfix = ".MECopy";
-        private static bool? isTimelineAvailable;
-        private static MethodInfo refreshInterpolablesList;
 
         /// <summary>
         /// Get a list of all the renderers of a GameObject
@@ -554,50 +552,6 @@ namespace MaterialEditorAPI
             Texture2D tex = bundle.LoadAsset<Texture2D>(assetPath);
             bundle.Unload(false);
             return tex;
-        }
-
-        private static System.Type GetKkapiType()
-        {
-#if KK
-            return System.Type.GetType("KKAPI.Utilities.TimelineCompatibility,KKAPI", throwOnError: false);
-#elif KKS
-            return System.Type.GetType("KKAPI.Utilities.TimelineCompatibility,KKSAPI", throwOnError: false);
-#elif AI
-            return System.Type.GetType("KKAPI.Utilities.TimelineCompatibility,AIAPI", throwOnError: false);
-#elif HS2
-            return System.Type.GetType("KKAPI.Utilities.TimelineCompatibility,HS2API", throwOnError: false);
-#else       //The others don't have timeline and/or studio, so no reason to check further
-            return null;
-#endif
-
-        }
-
-        internal static bool IsTimelineAvailable()
-        {
-            if (isTimelineAvailable == null)
-            {
-                var type = GetKkapiType();
-                if (type != null)
-                {
-                    var _isTimelineAvailable = type.GetMethod("IsTimelineAvailable", BindingFlags.Static | BindingFlags.Public);
-                    isTimelineAvailable = (bool)_isTimelineAvailable.Invoke(null, null);
-                }
-                else isTimelineAvailable = false;
-            }
-            return (bool)isTimelineAvailable;
-        }
-
-        internal static void RefreshInterpolablesList()
-        {
-            if (!(bool)isTimelineAvailable) return;
-
-            if (refreshInterpolablesList == null)
-            {
-                var type = GetKkapiType();
-                if (type != null)
-                    refreshInterpolablesList = type.GetMethod("RefreshInterpolablesList", BindingFlags.Static | BindingFlags.Public);
-            }
-            refreshInterpolablesList.Invoke(null, null);
         }
 
         /// <summary>
